+++ src/package/packageConvert.ts
@@ -112,6 +112,7 @@ export async function convertPackage(
       buildinstance: options.buildInstance,
       seedmetadata: options.seedMetadata,
       patchversion: options.patchversion,
+      codecoverage: options.codecoverage,
     },
     packageId,
     // TODO: createPackageVersionCreateRequest requires apiVersion exist.
@@ -129,15 +130,27 @@ export async function convertPackage(
 
   let results: Many<PackageVersionCreateRequestResult>;
   if (options.wait) {
-    results = await pollForStatusWithInterval(
-      createResult.id,
-      maxRetries,
-      packageId,
-      branch,
-      project,
-      connection,
-      options.frequency ?? Duration.seconds(POLL_INTERVAL_SECONDS)
-    );
+    try {
+      results = await pollForStatusWithInterval(
+        createResult.id,
+        maxRetries,
+        packageId,
+        branch,
+        project,
+        connection,
+        options.frequency ?? Duration.seconds(POLL_INTERVAL_SECONDS)
+      );
+    } catch (e) {
+      if (e instanceof SfError) {
+        const err = e as SfError;
+        if (err.name === 'PollingClientTimeout') {
+          err.setData({ VersionCreateRequestId: createResult.id });
+          err.message += ` Run 'sf package version create report -i ${createResult.id}' to check the status.`;
+        }
+        throw pkgUtils.applyErrorAction(pkgUtils.massageErrorMessage(err));
+      }
+      throw e;
+    }
   } else {
     results = await byId(packageId, connection);
   }
@@ -160,6 +173,7 @@ export async function createPackageVersionCreateRequest(
     buildinstance?: string;
     seedmetadata?: string;
     patchversion?: string;
+    codecoverage?: boolean;
   },
   packageId: string,
   apiVersion: string
@@ -243,7 +257,7 @@ export async function createPackageVersionCreateRequest(
 
 async function createRequestObject(
   packageId: string,
-  options: { installationkey?: string; buildinstance?: string },
+  options: { installationkey?: string; buildinstance?: string; codecoverage?: boolean },
   packageVersTmpRoot: string,
   packageVersBlobZipFile: string
 ): Promise<PackagingSObjects.Package2VersionCreateRequest> {
@@ -254,6 +268,7 @@ async function createRequestObject(
     InstallKey: options.installationkey,
     Instance: options.buildinstance,
     IsConversionRequest: true,
+    CalculateCodeCoverage: options.codecoverage ?? false,
   } as PackagingSObjects.Package2VersionCreateRequest;
   await fs.promises.rm(packageVersTmpRoot, { recursive: true });
   return requestObject;